refactor(skills): unify creation and editing under EditSkillPage - Use only the edit-skill route and select editing by skillId - Remove CreateSkillPage and update every library entry point - Verify creation, loading, retry, and row navigation behavior
